move modules tests to core/modules.rs
